+++ tests/integration/test_metrics.py
@@ -1163,7 +1163,6 @@ def test_no_transaction_metrics_when_filtered(mini_sentry, relay):
     ]
     filtered_events.sort(key=lambda x: x["category"])
 
-    # NOTE: span categories should be 2.
     assert filtered_events == [
         {"reason": "release-version", "category": "span", "quantity": 2},
         {"reason": "release-version", "category": "span_indexed", "quantity": 2},